requirements.txt
# Core dependencies
torch==2.4.1
torchvision==0.19.1
accelerate==1.0.1
diffusers==0.32.2
transformers==4.46.3
safetensors==0.5.3
peft==0.13.2
tokenizers==0.20.3
huggingface-hub==0.29.3

# Image processing
Pillow==10.4.0
numpy==1.24.4
opencv-python==4.11.0.86
scikit-image==0.21.0
einops==0.8.1

# API server
fastapi==0.115.11
uvicorn==0.33.0
python-multipart==0.0.20
pydantic==2.10.6

# Utilities
tqdm==4.67.1
requests==2.32.3
python-dotenv==1.0.1
tenacity==8.2.3
httpx==0.28.1

# Optional dependencies
gradio==4.44.1
matplotlib==3.7.5
av==12.3.0
fvcore==0.1.5.post20221221
iopath==0.1.10
omegaconf==2.3.0
cloudpickle==3.1.1
config==0.5.1